requirements.txt
asn1crypto==1.5.1
boto3==1.42.87
botocore==1.42.87
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
filelock==3.25.2
idna==3.11
jmespath==1.1.0
numpy==2.4.4
packaging==26.0
pandas==2.3.3
platformdirs==4.9.6
pyarrow==23.0.1
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.4.0
sortedcontainers==2.4.0
tomlkit==0.14.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
